I believe I have fixed the query repeatability problem, at the expense of some perf
Basically, UpdateByQuery is a cool feature but seems to be a scam that breaks when things get complicated. I replaced this with a separate search and update queries, with refresh set to true on the update, which now seems to have fixed everything. This would be faster if we could batch the updates, or you know just wait for them, but that would require a large change to a lot of the way the API works and that doesn't seem like a good use of time right now.